Return xs:decimal when dividing xs:long
Originally dividing xs:long was returning a double, but it should have
been returning a decimal instead. Similar fixes have been made for other
integer based types.

DAFFODIL-2063

+++ daffodil-core/src/main/scala/org/apache/daffodil/dpath/NodeInfoUtils.scala
@@ -86,9 +86,9 @@ object NodeInfoUtils {
       case Decimal => Decimal
       case Integer => Decimal
       case Double => Double
-      case Long => Double
+      case Long => Decimal
       case Float => Float
-      case Int => Float
+      case Int => Decimal
       case ArrayIndex => ArrayIndex
       case _ => Assert.usageError("Unsupported return type: %s".format(resultType))
     }
